need to think about database table structure User ,Room has ackward relationship

+++ src/main/java/nio_client/util/ElseProcess.java
@@ -116,7 +116,7 @@ public class ElseProcess
                                     }
                                     int notRoomRead = Integer.parseInt(s1[s1.length-2]);
                                     String usefulTime = s1[s1.length-1];
-                                    room.addNewTextToRoom((long)textId,sender,text,notRoomRead,usefulTime,room);
+                                    room.addNewTextToRoom(textId,sender,text,notRoomRead,usefulTime,room);
                                 });
                             } catch (IOException e)
                             {
